Move NmClip decompilation to a separate class.
